refactor: organise platoform_io JS logic into folders
