+++ nodes/web/helpers/platform_io/utils.js
@@ -1,5 +1,3 @@
-import { app } from "../../../scripts/app.js";
-import { addStyling } from "./platform_io_style.js";
 const NODES = {
   signature_input_image: "Input Image",
   signature_input_text: "Input Text",
@@ -9,7 +7,7 @@ const NODES = {
   signature_output: "Output",
 };
 
-function output(node, widget) {
+const output = (node, widget) => {
   const widgetType = widget.value.toUpperCase();
   if (node.inputs !== undefined) {
     for (const input of node.inputs) {
@@ -32,9 +30,9 @@ function output(node, widget) {
       break;
     }
   }
-}
+};
 
-function inputImage(node, widget) {
+const inputImage = (node, widget) => {
   const name = widget.value;
   const type = widget.value.toUpperCase();
   if (node.inputs !== undefined) {
@@ -49,9 +47,9 @@ function inputImage(node, widget) {
   } else {
     node.addOutput(name, type);
   }
-}
+};
 
-function inputNumber(node, widget) {
+const inputNumber = (node, widget) => {
   const widgetType = widget.value.toUpperCase();
   if (node.inputs !== undefined) {
     if (node.inputs.length > 0) {
@@ -88,9 +86,9 @@ function inputNumber(node, widget) {
       valueWidget.options.step = 0.01;
     }
   }
-}
+};
 
-function inputSelector(node, widget) {
+const inputSelector = (node, widget) => {
   const widgetType = widget.value.toUpperCase();
 
   const widgets = node.widgets || [];
@@ -101,9 +99,9 @@ function inputSelector(node, widget) {
   if (last_widget.value !== "") {
     node.addInput("input_" + widgets.length, widgetType);
   }
-}
+};
 
-function inputText(node, widget) {
+const inputText = (node, widget) => {
   // Safety check for widget
   if (!widget) {
     console.error("Widget is undefined in inputText");
@@ -122,7 +120,7 @@ function inputText(node, widget) {
     node.outputs[0].type = widgetType;
     node.outputs[0].name = widget.value;
   }
-}
+};
 
 const nodeWidgetHandlers = {
   signature_input_image: {
@@ -146,74 +144,13 @@ const nodeWidgetHandlers = {
 };
 
 // In the main function where widgetLogic is called
-function widgetLogic(node, widget) {
+const widgetLogic = (node, widget) => {
   // Retrieve the handler for the current node title and widget name
   const handler = nodeWidgetHandlers[node.comfyClass]?.[widget.name];
 
   if (handler) {
     handler(node, widget);
   }
-}
-
-const ext = {
-  name: "signature.platform_io",
-
-  nodeCreated(node) {
-    const title = node.comfyClass;
-
-    if (NODES.hasOwnProperty(title)) {
-      // Only set initial shape, don't try to style yet
-      node.shape = "box";
-
-      node.validateLinks = function () {
-        if (node.outputs !== undefined) {
-          if (node.outputs[0].type !== "*" && node.outputs[0].links) {
-            node.outputs[0].links
-              .filter((linkId) => {
-                const link = node.graph.links[linkId];
-                return link && link.type !== node.outputs[0].type && link.type !== "*";
-              })
-              .forEach((linkId) => {
-                node.graph.removeLink(linkId);
-              });
-          }
-        }
-      };
-
-      for (const w of node.widgets || []) {
-        let widgetValue = w.value;
-        widgetLogic(node, w);
-
-        // Check if the widget already has a value property descriptor
-        const descriptor = Object.getOwnPropertyDescriptor(w, "value");
-
-        // Only add our custom getter/setter if there isn't already one
-        if (!descriptor || (!descriptor.get && !descriptor.set)) {
-          // Store the actual value in a separate property
-          w._value = widgetValue;
-
-          try {
-            Object.defineProperty(w, "value", {
-              get() {
-                return w._value;
-              },
-              set(newVal) {
-                w._value = newVal;
-                widgetLogic(node, w);
-              },
-              configurable: true, // Allow property to be redefined if needed
-            });
-          } catch (error) {
-            console.warn("Could not define custom property for widget:", error);
-            // Fallback: just set the value and run widget logic
-            w.value = widgetValue;
-            widgetLogic(node, w);
-          }
-        }
-      }
-      addStyling(node, title);
-    }
-  },
 };
 
-app.registerExtension(ext);
+export { NODES, widgetLogic };
